rename files to remove um_* where not needed
